apps/llm/pyproject.toml
[tool.poetry]
name = "llmproxy"
version = "0.1.0"
description = ""
authors = []
packages = [{include = "src"}]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^3.0.3"
requests = "^2.32.3"
openai = "^1.52.0"
aiohttp = "^3.10.10"
python-dotenv = "^1.0.1"
flask-cors = "^5.0.0"
rich = "^13.9.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# CLI command configuration
[tool.poetry.scripts]
llm = "src.llm:run_llm"
